Переделал структуру проекта
