+++ gulpfile.js
@@ -12,6 +12,7 @@ const minify = require('gulp-csso');
 const rename = require('gulp-rename');
 const imagemin = require('gulp-imagemin');
 const rollup = require('gulp-better-rollup');
+const resolve = require('rollup-plugin-node-resolve');
 const sourcemaps = require('gulp-sourcemaps');
 const mocha = require('gulp-mocha');
 
@@ -42,7 +43,9 @@ gulp.task('scripts', function () {
   return gulp.src('js/main.js')
     .pipe(plumber())
     .pipe(sourcemaps.init())
-    .pipe(rollup({}, 'iife'))
+    .pipe(rollup({
+      plugins: [resolve()]
+     }, 'iife'))
     .pipe(sourcemaps.write(''))
     .pipe(gulp.dest('build/js'));
 });
