+++ js/components/gameThree/index.js
@@ -1,5 +1,5 @@
-import {getElementFromTemplate, changeView} from './util';
-import renderGameTwo from './game-2';
+import {getElementFromTemplate, changeView} from '../../util';
+import renderStats from '../stats';
 
 // Получаем documentFragment с dom-узлами из шаблона
 const documentFragmentBase = getElementFromTemplate(
@@ -19,29 +19,16 @@ const documentFragmentBase = getElementFromTemplate(
         </div>
       </header>
       <div class='game'>
-        <p class='game__task'>Угадайте для каждого изображения фото или рисунок?</p>
-        <form class='game__content'>
+        <p class='game__task'>Найдите рисунок среди изображений</p>
+        <form class='game__content  game__content--triple'>
           <div class='game__option'>
-            <img src='http://placehold.it/468x458' alt='Option 1' width='468' height='458'>
-            <label class='game__answer game__answer--photo'>
-              <input name='question1' type='radio' value='photo'>
-              <span>Фото</span>
-            </label>
-            <label class='game__answer game__answer--paint'>
-              <input name='question1' type='radio' value='paint'>
-              <span>Рисунок</span>
-            </label>
+            <img src='http://placehold.it/304x455' alt='Option 1' width='304' height='455'>
           </div>
           <div class='game__option'>
-            <img src='http://placehold.it/468x458' alt='Option 2' width='468' height='458'>
-            <label class='game__answer  game__answer--photo'>
-              <input name='question2' type='radio' value='photo'>
-              <span>Фото</span>
-            </label>
-            <label class='game__answer  game__answer--paint'>
-              <input name='question2' type='radio' value='paint'>
-              <span>Рисунок</span>
-            </label>
+            <img src='http://placehold.it/304x455' alt='Option 1' width='304' height='455'>
+          </div>
+          <div class='game__option'>
+            <img src='http://placehold.it/304x455' alt='Option 1' width='304' height='455'>
           </div>
         </form>
         <div class='stats'>
@@ -50,11 +37,11 @@ const documentFragmentBase = getElementFromTemplate(
             <li class='stats__result stats__result--slow'></li>
             <li class='stats__result stats__result--fast'></li>
             <li class='stats__result stats__result--correct'></li>
+            <li class='stats__result stats__result--wrong'></li>
             <li class='stats__result stats__result--unknown'></li>
+            <li class='stats__result stats__result--slow'></li>
             <li class='stats__result stats__result--unknown'></li>
-            <li class='stats__result stats__result--unknown'></li>
-            <li class='stats__result stats__result--unknown'></li>
-            <li class='stats__result stats__result--unknown'></li>
+            <li class='stats__result stats__result--fast'></li>
             <li class='stats__result stats__result--unknown'></li>
           </ul>
         </div>
@@ -66,24 +53,36 @@ export default () => {
   const documentFragment = documentFragmentBase.cloneNode(true);
 
   // Добавляем логику работы
-  const REQUIRED_ANSWERS_COUNT = 2;
   const content = documentFragment.querySelector(`.game__content`);
-  const answers = Array.from(documentFragment.querySelectorAll(`[type='radio']`));
+  const answers = Array.from(documentFragment.querySelectorAll(`.game__option`));
 
-  const getSelectedAnswers = () => {
-    const result = answers.filter(((answer) => {
-      return answer.checked;
-    }));
-
-    return result.length === REQUIRED_ANSWERS_COUNT;
+  const clearAnswersSelection = () => {
+    answers.forEach((answer) => {
+      answer.classList.remove(`game__option--selected`);
+    });
   };
 
-  content.addEventListener(`change`, () => {
+  const getSelectedAnswers = () => {
+    return answers.some(((answer) => {
+      return answer.classList.contains(`game__option--selected`);
+    }));
+  };
+
+  content.addEventListener(`click`, (e) => {
+    const option = e.target.closest(`.game__option`);
+
+    if (!option) {
+      return;
+    }
+
+    clearAnswersSelection();
+    option.classList.add(`game__option--selected`);
+
     if (!getSelectedAnswers()) {
       return;
     }
 
-    changeView(renderGameTwo());
+    changeView(renderStats());
   });
 
   // Возвращаем dom - элементы
